refactor: streamline feature initialization and modularize command logic
- Reorganized feature initialization in `main.ts` using a consistent `load()` function for better maintainability.
- Modularized `amethyst:lore` command into `features/commands/lore.ts` and consolidated command loading in `features/commands/index.ts`.
- Updated `loadWorldBorder` to accept `guild_id` for improved flexibility.
