+++ behaviour_pack/scripts-dev/features/commands/index.ts
@@ -1,0 +1,5 @@
+import loreCommand from "./lore";
+
+export default function loadCommands() {
+    loreCommand()
+}
